[IMP] project: add burndown chart in reporting menu
Before this commit we have reports that depict the status of a
project at a certain point in time, but we don't have any reports
representing the actual progress of the project in time.

This commit adds a burndown chart, this chart would help users see
the evolution of the project and determine whether it is on the
right track or not.

task-2458017