[IMP] project: add staked in line chart to create burndown chart
Before this commit, the stacked is not available in the line chart but
only in the bar chart.

This commit adds the "Stacked" button in the line chart and add the
possibility to stack in this line graph.

Additional features:

 - add same color for the background color when stacked

when we want to stacked in the line chart, the
background color is grey or black by default and we play with the
opacity of this color.

This commit takes the borderColor of each dataset and gives the same
color with opacity set to 0.4 for the background color when we want to
stack the line chart.

 - hide the pie chart button

task-2458017