[FIX] website_sale: not use t-if on attribute of snippet section
Before this commit the default product category was set on the product
snippet by a wrong use of qweb's <attribute> tag, trying to use t-if to
only apply the change for product snippet.
This actually was ignored and the attribute was appearing on all dropped
dynamic snippets (including blogs).

After this commit the default product category is set on the product
snippet from javascript.

https://github.com/odoo/odoo/pull/68529