[FIX] website_blog: fit the size of the Dexter effect darkening area
Before this commit the darkening area of the Dexter effect is visible
through the rounded corners when the blog post was not hovered.

After this commit the darkening area is moved inside the right element
thus not requiring magic numbers for truncating the area and fixing the
display of the darkening area behind the rounded corners.

closes odoo/odoo#68640

X-original-commit: ac69deabf7f773974523a9a8216c39c239d33ca7
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>