[FIX] project: keep create enabled in dependent task
Prior to this commit:

Once the user clicked on View Task in the "Blocked By" task list, the
Create button in the form view of the targetted task is hidden. Also,
in the subtask notebook the user is not able to create additionnal
subtasks.

Reason:

In the action, the create argument is set to False.

After this commit:

We merge the two action_open_task and action_open_task_form into a
unique one. We remove the create set to false.

Related PR : #66740

closes odoo/odoo#68652

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>